MPI_Request used to store request handles
The routine was using ints to hold request handles in MPI_Irecv and
MPI_Wait calls. I have rewritten the code so that they are of type
MPI_Request, as required by the MPI standard. This may not cause any
problems if you're using an MPICH-based MPI implementation because
MPI_Request is implemented as an int. However, OpenMPI implements
MPI_Request as a pointer to a structure so the compiler will flag a type
mismatch if MPI_Request is not being used. I also replaced the loop over
the MPI_Wait statements by a single MPI_Waitall, given I was editing
those lines anyway.